feat(scvhedules): implement new embed storage system with schedules
