+++ src/interaction/subcommands/roles/reactionRoles/reactionRolesAdd.ts
@@ -1,33 +1,31 @@
 import Modules from '@/constants/bot/commands/Modules';
-import Limits from '@/constants/database/Limits';
 import { Auxdibot } from '@/interfaces/Auxdibot';
 import { GuildAuxdibotCommandData } from '@/interfaces/commands/AuxdibotCommandData';
 import AuxdibotCommandInteraction from '@/interfaces/commands/AuxdibotCommandInteraction';
 import { AuxdibotSubcommand } from '@/interfaces/commands/AuxdibotSubcommand';
-import addReactionRole from '@/modules/features/roles/reaction_roles/addReactionRole';
+import argumentsToEmbedParameters from '@/util/argumentsToEmbedParameters';
 import handleError from '@/util/handleError';
-import handleLog from '@/util/handleLog';
-import { testLimit } from '@/util/testLimit';
+import parsePlaceholders from '@/util/parsePlaceholder';
 import { EmbedBuilder } from '@discordjs/builders';
-import { LogAction, ReactionRoleType } from '@prisma/client';
 import { ChannelType } from 'discord.js';
+import addReactionRole from '@/modules/features/roles/reaction_roles/addReactionRole';
+import { testLimit } from '@/util/testLimit';
+import Limits from '@/constants/database/Limits';
+import { ReactionRoleType } from '@prisma/client';
 
 export const reactionRolesAdd = <AuxdibotSubcommand>{
    name: 'add',
    info: {
       module: Modules['Roles'],
-      description: 'Add a reaction role to the server.',
-      usageExample: '/reaction_roles add (channel) (roles) [type]',
+      description: 'Add a reaction role to the server with custom Embed parameters.',
+      usageExample: '/reaction_roles add (channel) (roles) [type] (id)',
    },
    async execute(auxdibot: Auxdibot, interaction: AuxdibotCommandInteraction<GuildAuxdibotCommandData>) {
       if (!interaction.data) return;
       const channel = interaction.options.getChannel('channel', true, [ChannelType.GuildText]),
          roles = interaction.options.getString('roles', true),
-         title = interaction.options.getString('title') || 'React to receive roles!',
-         type = interaction.options.getString('type', false) || 'DEFAULT',
-         webhook_url = interaction.options.getString('webhook_url');
-      const split = roles.split(' ');
-      const builder = [];
+         id = interaction.options.getString('id', true),
+         type = interaction.options.getString('type', false) || 'DEFAULT';
       if (!testLimit(interaction.data.guildData.reaction_roles, Limits.REACTION_ROLE_DEFAULT_LIMIT)) {
          return await handleError(
             auxdibot,
@@ -39,6 +37,12 @@ export const reactionRolesAdd = <AuxdibotSubcommand>{
       if (!ReactionRoleType[type]) {
          return await handleError(auxdibot, 'INVALID_TYPE', 'This is not a valid reaction role type!', interaction);
       }
+      const stored = interaction.data.guildData.stored_embeds.find((i) => i.id === id);
+      if (!stored) return handleError(auxdibot, 'EMBED_NOT_FOUND', 'Embed not found!', interaction);
+      const { content, embed, webhook_url } = stored;
+
+      const split = roles.split(' ');
+      const builder = [];
       while (split.length) {
          const [emoji, roleID] = split.splice(0, 2);
          builder.push({ emoji, roleID });
@@ -51,14 +55,25 @@ export const reactionRolesAdd = <AuxdibotSubcommand>{
             interaction,
          );
       }
+      const parameters = argumentsToEmbedParameters(interaction);
       addReactionRole(
          auxdibot,
          interaction.guild,
          channel,
-         title,
+         parameters.title,
          builder,
-         undefined,
-         undefined,
+         JSON.parse(
+            await parsePlaceholders(auxdibot, JSON.stringify(embed), {
+               guild: interaction.data.guild,
+               member: interaction.data.member,
+            }),
+         ),
+         content
+            ? await parsePlaceholders(auxdibot, JSON.stringify(content), {
+                 guild: interaction.data.guild,
+                 member: interaction.data.member,
+              })
+            : undefined,
          ReactionRoleType[type],
          webhook_url,
       )
@@ -66,19 +81,13 @@ export const reactionRolesAdd = <AuxdibotSubcommand>{
             const resEmbed = new EmbedBuilder().setColor(auxdibot.colors.accept).toJSON();
             resEmbed.title = '👈 Created Reaction Role';
             resEmbed.description = `Created a reaction role in ${channel}`;
-            handleLog(auxdibot, interaction.data.guild, {
-               userID: interaction.data.member.id,
-               description: `Created a reaction role in ${channel.name}`,
-               type: LogAction.REACTION_ROLE_ADDED,
-               date: new Date(),
-            });
             return await auxdibot.createReply(interaction, { embeds: [resEmbed] });
          })
-         .catch((x) => {
+         .catch(async (x) => {
             handleError(
                auxdibot,
                'REACTION_ROLE_CREATE_ERROR',
-               typeof x.message == 'string' ? x.message : "Couldn't create that reaction role!",
+               typeof x.message == 'string' ? x.message : "Couldn't delete that reaction role!",
                interaction,
             );
          });
